test constants to own file
